playerstate refactor start
