+++ core/src/main/java/smashdudes/ecs/components/PlayerComponent.java
@@ -2,6 +2,7 @@ package smashdudes.ecs.components;
 
 import smashdudes.core.PlayerHandle;
 import smashdudes.ecs.Component;
+import smashdudes.gameplay.PlayerState;
 
 public class PlayerComponent extends Component
 {
@@ -11,6 +12,7 @@ public class PlayerComponent extends Component
 
     public boolean facingLeft = false;
 
+    public PlayerState currentState = PlayerState.Air_Idle;
     public PlayerComponent(PlayerHandle handle, String name)
     {
         this.handle = handle;
