--- core/src/main/java/smashdudes/ecs/components/PlayerIdleComponent.java
@@ -1,7 +1,0 @@
-package smashdudes.ecs.components;
-
-import smashdudes.ecs.Component;
-
-public class PlayerIdleComponent extends Component
-{
-}
