--- core/src/main/java/smashdudes/ecs/components/PlayerRunningComponent.java
@@ -1,7 +1,0 @@
-package smashdudes.ecs.components;
-
-import smashdudes.ecs.Component;
-
-public class PlayerRunningComponent extends Component
-{
-}
