+++ core/src/main/java/smashdudes/gameplay/PlayerState.java
@@ -1,0 +1,13 @@
+package smashdudes.gameplay;
+
+public enum PlayerState
+{
+    Ground_Idle,
+    Ground_Attack,
+    Ground_Stunned,
+    Ground_Running,
+
+    Air_Idle,
+}
+
+
